+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.8.0
 
   * if a Device changes Port, the Switch of old and new Port are commited automatically
+  * Settings panel is now able to remove all Devices without a description
   * moved installer from fabric to ansible
 
 ## v0.7.0